Remove free + used space check
This can also be solved by using `info.free` instead of `info.available`, though this check isn't necessary as there is a free space check before accepting new shards to prevent disk "no space left on device" errors and crashing.

There is a case where `available` will be missing 5% as it is reservered for root user that can lead to an issue where on startup the amount allocated would give an error. This is technically not incorrect.